requirements.txt
PyGitHub>=1.25